move tests to dk.mada
