+++ src/test/java/dk/mada/unit/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Action unit tests.
  */
-package dk.unit;
+package dk.mada.unit;
